chore: Move protocol id to xxi node crate
