+++ coordinator/src/dlc_protocol.rs
@@ -14,93 +14,16 @@ use diesel::Connection;
 use diesel::PgConnection;
 use diesel::QueryResult;
 use dlc_manager::ContractId;
-use dlc_manager::ReferenceId;
 use rust_decimal::prelude::FromPrimitive;
 use rust_decimal::prelude::ToPrimitive;
 use rust_decimal::Decimal;
-use std::fmt::Display;
-use std::fmt::Formatter;
-use std::str::from_utf8;
 use time::OffsetDateTime;
 use tokio::sync::broadcast::Sender;
-use uuid::Uuid;
 use xxi_node::cfd::calculate_pnl;
 use xxi_node::commons;
 use xxi_node::commons::Direction;
 use xxi_node::node::rust_dlc_manager::DlcChannelId;
-
-#[derive(Debug, Copy, Clone, PartialEq)]
-pub struct ProtocolId(Uuid);
-
-impl ProtocolId {
-    pub fn new() -> Self {
-        ProtocolId(Uuid::new_v4())
-    }
-
-    pub fn to_uuid(&self) -> Uuid {
-        self.0
-    }
-}
-
-impl Default for ProtocolId {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl Display for ProtocolId {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        self.0.to_string().fmt(f)
-    }
-}
-
-impl From<ProtocolId> for ReferenceId {
-    fn from(value: ProtocolId) -> Self {
-        let uuid = value.to_uuid();
-
-        // 16 bytes.
-        let uuid_bytes = uuid.as_bytes();
-
-        // 32-digit hex string.
-        let hex = hex::encode(uuid_bytes);
-
-        // Derived `ReferenceId`: 32-bytes.
-        let hex_bytes = hex.as_bytes();
-
-        let mut array = [0u8; 32];
-        array.copy_from_slice(hex_bytes);
-
-        array
-    }
-}
-
-impl TryFrom<ReferenceId> for ProtocolId {
-    type Error = anyhow::Error;
-
-    fn try_from(value: ReferenceId) -> Result<Self> {
-        // 32-digit hex string.
-        let hex = from_utf8(&value)?;
-
-        // 16 bytes.
-        let uuid_bytes = hex::decode(hex)?;
-
-        let uuid = Uuid::from_slice(&uuid_bytes)?;
-
-        Ok(ProtocolId(uuid))
-    }
-}
-
-impl From<Uuid> for ProtocolId {
-    fn from(value: Uuid) -> Self {
-        ProtocolId(value)
-    }
-}
-
-impl From<ProtocolId> for Uuid {
-    fn from(value: ProtocolId) -> Self {
-        value.0
-    }
-}
+use xxi_node::node::ProtocolId;
 
 pub struct DlcProtocol {
     pub id: ProtocolId,
@@ -605,8 +528,7 @@ impl DlcProtocolExecutor {
         Ok(())
     }
 
-    /// Completes the rollover dlc protocol as successful and updates the 10101 meta data
-    /// accordingly in a single database transaction.
+    /// Completes the collab close dlc protocol as successful
     fn finish_close_channel_dlc_protocol(
         &self,
         conn: &mut PgConnection,
