+++ coordinator/src/trade/mod.rs
@@ -3,7 +3,6 @@ use crate::db;
 use crate::decimal_from_f32;
 use crate::dlc_protocol;
 use crate::dlc_protocol::DlcProtocolType;
-use crate::dlc_protocol::ProtocolId;
 use crate::message::OrderbookMessage;
 use crate::node::Node;
 use crate::orderbook::db::matches;
@@ -52,6 +51,7 @@ use xxi_node::commons::TradeAndChannelParams;
 use xxi_node::commons::TradeParams;
 use xxi_node::node::event::NodeEvent;
 use xxi_node::node::signed_channel_state_name;
+use xxi_node::node::ProtocolId;
 
 pub mod models;
 pub mod websocket;
@@ -369,7 +369,7 @@ impl TradeExecutor {
                 trade_params.filled_with.clone(),
                 contract_input,
                 trade_params.pubkey,
-                protocol_id.into(),
+                protocol_id,
             )
             .await
             .context("Could not propose DLC channel")?;
@@ -548,7 +548,7 @@ impl TradeExecutor {
                 trade_params.filled_with.clone(),
                 &dlc_channel_id,
                 contract_input,
-                protocol_id.into(),
+                protocol_id,
             )
             .await
             .context("Could not propose reopen DLC channel update")?;
@@ -721,7 +721,7 @@ impl TradeExecutor {
                 trade_params.filled_with.clone(),
                 &dlc_channel_id,
                 contract_input,
-                protocol_id.into(),
+                protocol_id,
             )
             .await
             .context("Could not propose resize DLC channel update")?;
@@ -878,7 +878,7 @@ impl TradeExecutor {
                 trade_params.filled_with.clone(),
                 &channel_id,
                 settlement_amount_trader,
-                protocol_id.into(),
+                protocol_id,
             )
             .await?;
