+++ crates/xxi-node/src/node/mod.rs
@@ -36,6 +36,7 @@ use std::fmt::Display;
 use std::fmt::Formatter;
 use std::net::SocketAddr;
 use std::path::Path;
+use std::str::from_utf8;
 use std::sync::mpsc;
 use std::sync::Arc;
 use std::time::Duration;
@@ -428,15 +429,75 @@ impl Display for NodeInfo {
     }
 }
 
-pub fn new_reference_id() -> ReferenceId {
-    let uuid = Uuid::new_v4();
-    let hex = hex::encode(uuid.as_simple().as_ref());
-    let bytes = hex.as_bytes();
+#[derive(Debug, Copy, Clone, PartialEq)]
+pub struct ProtocolId(Uuid);
 
-    debug_assert!(bytes.len() == 32, "length must be exactly 32 bytes");
+impl ProtocolId {
+    pub fn new() -> Self {
+        ProtocolId(Uuid::new_v4())
+    }
 
-    let mut array = [0u8; 32];
-    array.copy_from_slice(bytes);
+    pub fn to_uuid(&self) -> Uuid {
+        self.0
+    }
+}
 
-    array
+impl Default for ProtocolId {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+impl Display for ProtocolId {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        self.0.to_string().fmt(f)
+    }
+}
+
+impl From<ProtocolId> for ReferenceId {
+    fn from(value: ProtocolId) -> Self {
+        let uuid = value.to_uuid();
+
+        // 16 bytes.
+        let uuid_bytes = uuid.as_bytes();
+
+        // 32-digit hex string.
+        let hex = hex::encode(uuid_bytes);
+
+        // Derived `ReferenceId`: 32-bytes.
+        let hex_bytes = hex.as_bytes();
+
+        let mut array = [0u8; 32];
+        array.copy_from_slice(hex_bytes);
+
+        array
+    }
+}
+
+impl TryFrom<ReferenceId> for ProtocolId {
+    type Error = anyhow::Error;
+
+    fn try_from(value: ReferenceId) -> Result<Self> {
+        // 32-digit hex string.
+        let hex = from_utf8(&value)?;
+
+        // 16 bytes.
+        let uuid_bytes = hex::decode(hex)?;
+
+        let uuid = Uuid::from_slice(&uuid_bytes)?;
+
+        Ok(ProtocolId(uuid))
+    }
+}
+
+impl From<Uuid> for ProtocolId {
+    fn from(value: Uuid) -> Self {
+        ProtocolId(value)
+    }
+}
+
+impl From<ProtocolId> for Uuid {
+    fn from(value: ProtocolId) -> Self {
+        value.0
+    }
 }
